3h48, hope u sleep well

+++ FINAL COMPILED/abc/src/main/java/com/abc/backend/CNPM/service/NhaCungCapLuuTru.java
@@ -18,5 +18,5 @@ public interface NhaCungCapLuuTru {
 
     Vehicle luuXe(Vehicle xe);
 
-    void xoaXe(Long id);
+    void xoaXe(Integer id);
 }
